Removed the packaging code from the build system

+++ code/custom/languages/4coder_cpp_lexer_test.cpp
@@ -436,7 +436,7 @@ print_token_list(Token_List *list, String_Const_u8 text){
     for (Token_Block *block = list->first;
          block != 0;
          block = block->next){
-        i32 count = block->count;
+        i64 count = block->count;
         Token *token = block->tokens;
         for (i32 i = 0; i < count; i += 1, token += 1){
             printf("[%5llu, %5llu) %20s / %20s : 0x%04x / 0x%04x\n",
@@ -456,7 +456,7 @@ file_read_all(Arena *arena, FILE *file){
     result.size = ftell(file);
     fseek(file, 0, SEEK_SET);
     result.str = push_array(arena, u8, result.size + 1);
-    fread(result.str, result.size, 1, file);
+    fread(result.str, (size_t)result.size, 1, file);
     result.str[result.size] = 0;
     return(result);
 }
