+++ code/custom/lexer_generator/4coder_lex_gen_main.cpp
@@ -3670,7 +3670,7 @@ gen_contiguous_control_flow_lexer(Arena *scratch, Token_Kind_Set tokens, Lexer_M
          flag = flag->next){
         Flag_Reset_Rule reset_rule = flag->reset_rule;
         Flag_Bind_Property bind_property =
-            (flag->emit_flags != 0)?FlagBindProperty_Bound:FlagBindProperty_Free;
+        (flag->emit_flags != 0)?FlagBindProperty_Bound:FlagBindProperty_Free;
         
         Flag_Bucket *bucket = &bucket_set.buckets[bind_property][reset_rule];
         Flag_Ptr_Node *node = push_array(scratch, Flag_Ptr_Node, 1);
@@ -3900,7 +3900,7 @@ file_read_all(Arena *arena, FILE *file){
     result.size = ftell(file);
     fseek(file, 0, SEEK_SET);
     result.str = push_array(arena, u8, result.size + 1);
-    fread(result.str, result.size, 1, file);
+    fread(result.str, (size_t)result.size, 1, file);
     result.str[result.size] = 0;
     return(result);
 }
